Move capturePage to WebContents
