+++ lib/browser/api/browser-window.js
@@ -153,6 +153,9 @@ Object.assign(BrowserWindow.prototype, {
   },
   showDefinitionForSelection () {
     return this.webContents.showDefinitionForSelection()
+  },
+  capturePage (...args) {
+    return this.webContents.capturePage.apply(this.webContents, args)
   }
 })
