+++ lib/renderer/web-view/web-view.js
@@ -386,7 +386,8 @@ var registerWebViewElement = function () {
     'inspectServiceWorker',
     'print',
     'printToPDF',
-    'showDefinitionForSelection'
+    'showDefinitionForSelection',
+    'capturePage'
   ]
   nonblockMethods = [
     'insertCSS',
